requirements.txt
absl-py==1.4.0
asttokens==2.2.1
backcall==0.2.0
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
cmake==3.27.2
colorama==0.4.6
contourpy==1.1.0
cycler==0.11.0
decorator==5.1.1
executing==1.2.0
filelock==3.12.2
fonttools==4.42.1
google-auth==2.22.0
google-auth-oauthlib==1.0.0
grpcio==1.57.0
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.0.1
ipython==8.12.2
jedi==0.19.0
jinja2==3.1.2
kiwisolver==1.4.4
lit==16.0.6
markdown==3.4.4
markupsafe==2.1.3
matplotlib==3.7.2
matplotlib-inline==0.1.6
mpmath==1.3.0
networkx==3.1
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
numpy==1.23.5
oauthlib==3.2.2
opencv-python-headless
packaging==23.1
pandas==2.0.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
pillow==10.0.0
prompt-toolkit==3.0.39
protobuf==4.21.2
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pygments==2.16.1
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
pyyaml==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scipy==1.10.1
seaborn==0.12.2
six==1.16.0
stack-data==0.6.2
sympy==1.12
tensorboard==2.14.0
tensorboard-data-server==0.7.1
thop==0.1.1-2209072238
torch==2.0.1
torchaudio==2.0.2
torchvision==0.15.2
tqdm==4.66.1
traitlets==5.9.0
triton==2.0.0
typing-extensions==4.7.1
tzdata==2023.3
urllib3==1.26.16
wcwidth==0.2.6
werkzeug==2.3.7
zipp==3.16.2
